compare userId with product's userId logic

+++ src/main/java/com/example/db/controller/ProductController.java
@@ -47,26 +47,24 @@ public class ProductController {
     @PutMapping("/{productId}")
     public ResponseEntity<?> updateProduct(@PathVariable Long productId, @RequestBody Product product, Authentication authentication) {
         //productId에 userId 검증 추가
-        return ResponseEntity.ok(productService.updateProduct(product));
 
-//        try{
-//            Long userId = (Long)authentication.getPrincipal();
-//            if(productService.isProductOwnedByUser(productId,userId)){
-//                return ResponseEntity.ok(productService.updateProduct(product));
-//            }
-//
-//        }catch(Exception e){
-//            log.error("product's userId is not match user's id");
-//        }
-//        return ResponseEntity.ok("update not complete");
+        try{
+            Long userId = (Long)authentication.getPrincipal();
+            if(productService.isProductOwnedByUser(productId,userId)){
+                return ResponseEntity.ok(productService.updateProduct(product));
+            }
+
+        }catch(Exception e){
+            log.error("product's userId is not match user's id");
+        }
+        return ResponseEntity.ok("update not complete");
     }
     @DeleteMapping("/{productId}")
     public ResponseEntity<?> deleteProduct(@PathVariable Long productId,Authentication authentication){
-        //Long id = (Long)authentication.getPrincipal();
-        productService.deleteProduct(productId);
-//        if(productService.isProductOwnedByUser(productId,id)){
-//            productService.deleteProduct(productId);
-//        }
+        Long id = (Long)authentication.getPrincipal();
+        if(productService.isProductOwnedByUser(productId,id)){
+            productService.deleteProduct(productId);
+        }
         return ResponseEntity.ok("delete complete successfully!");
     }
